refactor(balancer) split dns and balancer logic
A lot of the code in the balancer is DNS related, fanning out hosts
to underlying addresses and tracking them over DNS updates.
This disentangles them, such that the DNS related code now becomes
a base class that can be used to build balancers on top. The ring
balancer just being one of them.
